gps-producer/requirements.txt
kafka-python==2.0.2
faker==25.6.0